+++ dst/app/controllers/api/performances/suspended.js
@@ -26,10 +26,11 @@ const moment = require("moment-timezone");
 const numeral = require("numeral");
 const debug = createDebug('ttts-staff:controllers');
 const EMPTY_STRING = '-';
-const EV_SERVICE_STATUS_NAMES = {};
-EV_SERVICE_STATUS_NAMES[tttsapi.factory.performance.EvServiceStatus.Normal] = EMPTY_STRING;
-EV_SERVICE_STATUS_NAMES[tttsapi.factory.performance.EvServiceStatus.Slowdown] = '一時休止';
-EV_SERVICE_STATUS_NAMES[tttsapi.factory.performance.EvServiceStatus.Suspended] = '完全中止';
+// const EV_SERVICE_STATUS_NAMES: any = {
+// };
+// EV_SERVICE_STATUS_NAMES[tttsapi.factory.performance.EvServiceStatus.Normal] = EMPTY_STRING;
+// EV_SERVICE_STATUS_NAMES[tttsapi.factory.performance.EvServiceStatus.Slowdown] = '一時休止';
+// EV_SERVICE_STATUS_NAMES[tttsapi.factory.performance.EvServiceStatus.Suspended] = '完全中止';
 const REFUND_STATUS_NAMES = {};
 REFUND_STATUS_NAMES[tttsapi.factory.performance.RefundStatus.None] = EMPTY_STRING;
 REFUND_STATUS_NAMES[tttsapi.factory.performance.RefundStatus.NotInstructed] = '未指示';
@@ -165,13 +166,16 @@ function findSuspendedPerformances(req, conditions) {
                 }
             }
             const tourNumber = (_c = (_b = performance.additionalProperty) === null || _b === void 0 ? void 0 : _b.find((p) => p.name === 'tourNumber')) === null || _c === void 0 ? void 0 : _c.value;
-            let evServiceStatus = tttsapi.factory.performance.EvServiceStatus.Normal;
+            // let evServiceStatus = tttsapi.factory.performance.EvServiceStatus.Normal;
+            let evServiceStatusName = EMPTY_STRING;
             switch (performance.eventStatus) {
                 case cinerinoapi.factory.chevre.eventStatusType.EventCancelled:
-                    evServiceStatus = tttsapi.factory.performance.EvServiceStatus.Suspended;
+                    // evServiceStatus = tttsapi.factory.performance.EvServiceStatus.Suspended;
+                    evServiceStatusName = '完全中止';
                     break;
                 case cinerinoapi.factory.chevre.eventStatusType.EventPostponed:
-                    evServiceStatus = tttsapi.factory.performance.EvServiceStatus.Slowdown;
+                    // evServiceStatus = tttsapi.factory.performance.EvServiceStatus.Slowdown;
+                    evServiceStatusName = '一時休止';
                     break;
                 case cinerinoapi.factory.chevre.eventStatusType.EventScheduled:
                     break;
@@ -185,10 +189,8 @@ function findSuspendedPerformances(req, conditions) {
                 start_date: performance.startDate,
                 end_date: performance.endDate,
                 tour_number: tourNumber,
-                ev_service_status: evServiceStatus,
-                ev_service_status_name: (evServiceStatus !== undefined)
-                    ? EV_SERVICE_STATUS_NAMES[evServiceStatus]
-                    : undefined,
+                // ev_service_status: evServiceStatus,
+                ev_service_status_name: evServiceStatusName,
                 online_sales_update_at: extension === null || extension === void 0 ? void 0 : extension.online_sales_update_at,
                 online_sales_update_user: extension === null || extension === void 0 ? void 0 : extension.online_sales_update_user,
                 canceled: numberOfReservations,